First pass at abstraction and database changelog update.
