+++ service/build.gradle
@@ -80,7 +80,7 @@ dependencies {
 // Changed this to enable inline development of the includes listed below.
 def includes = [
   'web-toolkit-ce'   : 'com.k_int.grails:web-toolkit-ce:3.4.4',
-  'grails-okapi'  : 'com.k_int.okapi:grails-okapi:3.5.3'
+  'grails-okapi'  : 'com.k_int.okapi:grails-okapi:3.5.5'
 ]
 
 bootRun {
