+++ service/grails-app/conf/application.yml
@@ -133,13 +133,12 @@ environments:
       gorm:
         multiTenancy:
           mode: SCHEMA
-          tenantResolverClass: org.grails.datastore.mapping.multitenancy.resolvers.SystemPropertyTenantResolver # System prop of gorm.tenantId
+          tenantResolverClass: org.grails.datastore.mapping.multitenancy.resolvers.SystemPropertyTenantResolver # System prop of gorm.tenantId 
       plugin:
         databasemigration:
-          updateOnStart: true
-          updateOnStartFileName: module-tenant-changelog.groovy
+          changelogFileName: module-tenant-changelog.groovy
     gorm:
-      tenantId: public # Used by SystemPropertyTenantResolver
+      tenantId: database_test # Used by SystemPropertyTenantResolver
 
   production:
     dataSource:
