+++ service/grails-app/domain/org/olf/licenses/LicenseOrg.groovy
@@ -19,7 +19,7 @@ public class LicenseOrg implements MultiTenant<LicenseOrg>{
   RefdataValue role
 
   static belongsTo = [
-    owner: License
+    owner: LicenseAmendment
   ]
 
     static mapping = {
